[1.0.0] Remove deprecated methods/code
In preparation for `v1.0.0`, remove all deprecated code.
