+++ lib/timezone/error.rb
@@ -12,16 +12,12 @@ module Timezone
     class Base < StandardError; end
     # Indicates an invalid timezone name.
     class InvalidZone < Base; end
-    # @deprecated this class will be removed in the next release of the gem.
-    class NilZone < Base; end
     # Indicates a lookup failure.
     class Lookup < Base; end
     # Indicates an error during lookup using the geonames API.
     class GeoNames < Lookup; end
     # Indicates an error during lookup using the google API.
     class Google < Lookup; end
-    # @deprecated this class will be removed in the next release of the gem.
-    class ParseTime < Base; end
     # Indicates a missing stub during a test lookup.
     class Test < Lookup; end
     # Indicates an invalid configuration.
